# bin/Debug/net10.0/OpenAPIDoc.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FormulariosAPI | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5023/"
    }
  ],
  "paths": {
    "/api/Areas": {
      "get": {
        "tags": [
          "Areas"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AreaDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AreaDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AreaDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Areas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AreaCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AreaCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Areas/{id}": {
      "get": {
        "tags": [
          "Areas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Areas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Cargas": {
      "get": {
        "tags": [
          "Cargas"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargaDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargaDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargaDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cargas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargaCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CargaCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CargaCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CargaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cargas/area/{areaId}": {
      "get": {
        "tags": [
          "Cargas"
        ],
        "parameters": [
          {
            "name": "areaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargaDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargaDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargaDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Cargas/{id}": {
      "get": {
        "tags": [
          "Cargas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CargaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargaDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Forms": {
      "get": {
        "tags": [
          "Forms"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Forms"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FormCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Forms/area/{areaId}": {
      "get": {
        "tags": [
          "Forms"
        ],
        "parameters": [
          {
            "name": "areaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Forms/{id}": {
      "get": {
        "tags": [
          "Forms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Forms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Forms/{id}/publish": {
      "patch": {
        "tags": [
          "Forms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AreaCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "AreaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CargaCreateDto": {
        "type": "object",
        "properties": {
          "formId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "areaId": {
            "type": "string",
            "format": "uuid"
          },
          "dataJson": {
            "type": "string"
          }
        }
      },
      "CargaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "formId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "areaId": {
            "type": "string",
            "format": "uuid"
          },
          "dataJson": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FormCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "schemaJson": {
            "type": "string"
          },
          "areaId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "FormDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "schemaJson": {
            "type": "string"
          },
          "isPublished": {
            "type": "boolean"
          },
          "areaId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "areaId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Areas"
    },
    {
      "name": "Auth"
    },
    {
      "name": "Cargas"
    },
    {
      "name": "Forms"
    }
  ]
}